Applying RefreshableJpaRepositoryImpl in the app init
This commit introduces the annotation @EnableJpaRepositories with the parameter repositoryBaseClass set to RefreshableJpaRepositoryImpl.class. By configuring this, the application is now empowered to utilize the RefreshableJpaRepository implementation as the base class for JPA repositories. This change enhances the refreshability and reliability of JPA repositories, contributing to a more robust data management system.