Improve Exception Handling in Producer/Studio Repo
Updated the error handling in the ProducerRepository and StudioRepository classes to provide more informative error messages when entities are not found by ID. Previously, the repositories used a simple orElseThrow method, which did not provide specific details about the missing entity. Now, the repositories use the orElseThrow method with a custom exception message, indicating the ID of the entity that was not found.
Additionally, a new exception handler has been added to the ErrorHandler.java file to handle the custom exception InvalidBeanFromCsvException. This handler returns a badRequest response with a detailed error message in case such an exception occurs during CSV processing. This enhancement improves the error reporting mechanism and provides better feedback to users in case of entity lookup failures.