Add Docker Compose for Development Environment
Added a Docker Compose file (docker-compose.yml) to facilitate setting up the development environment. The Docker Compose configuration defines a service named gra-api based on the Maven 3.9.6 image with Eclipse Temurin 17 running on Ubuntu Focal. The service mounts the project directory as a volume inside the container and sets the working directory to the project directory. The command specified in the service builds the project using Maven (mvn clean package) and then runs the generated JAR file (java -jar ./target/golden-raspberry-awards-api-0.0.1-SNAPSHOT.jar). Additionally, the service maps port 8080 of the container to port 8080 of the host machine, allowing access to the API from localhost. This Docker Compose configuration streamlines the setup process for local development, providing a consistent and isolated environment for running the application.